+++ InstaAddict/__init__.py
@@ -1,6 +1,6 @@
 """InstaAddict - Human-like Instagram bot powered by UIAutomator2"""
 
-__version__ = "1.4.0"
+__version__ = "1.4.1"
 __tested_ig_version__ = "447.0.0.55.81"
 
 import os